pyproject.toml
[project]
name = "receipt-organizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "img2pdf>=0.6.3",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
]
